refactor(validator): decompose main.py into focused modules (ORO-838)
main.py was 1176 lines doing everything: argparse, bittensor setup, the
run loop, agent download + sandbox exec, log upload, completion reporting,
Watchtower auto-update polling, and inference-token validation. Hard to test
or modify any one piece without touching the rest.

Split into:
* config.py (build_config, configure_logging, METRICS_PORT)
* sandbox_runner.py (SandboxRunner: download_agent, run_sandbox)
* completion.py (CompletionReporter: upload_logs, complete_run, complete_with_failure)
* auto_update.py (Watchtower poller + sandbox image pull)
* inference_token.py (validate_inference_token, validation_model_for)
* url_utils.py (rewrite_localhost_url, shared by sandbox_runner + completion)

main.py drops to 538 lines and only hosts the orchestration that needs
Validator state (run loop, run_evaluation_cycle, fetch_problems,
setup_bittensor_objects).

Pure structural refactor — no functional changes. 389 tests pass.

Refreshed test_auto_update.py + test_validate_inference_token.py to target
the new module-level functions instead of the removed `Validator._private`
methods.
