+++ subnet/tests/validator/test_auto_update.py
@@ -1,18 +1,11 @@
-"""Tests for _check_for_updates auto-update logic."""
+"""Tests for validator.auto_update.check_for_updates."""
 
 from unittest.mock import MagicMock, patch
 
 import pytest
 import requests
 
-import validator.main as main_module
-from validator.main import Validator
-
-
-@pytest.fixture
-def validator_instance():
-    with patch.object(Validator, "__init__", lambda self: None):
-        yield Validator()
+from validator import auto_update
 
 
 def _mock_get(watchtower_resp=MagicMock(ok=True, status_code=200), health_ok=True):
@@ -34,23 +27,28 @@ def _mock_get(watchtower_resp=MagicMock(ok=True, status_code=200), health_ok=Tru
 
 
 class TestCheckForUpdates:
-    def test_skipped_when_disabled(self, validator_instance):
+    def test_skipped_when_disabled(self):
         with (
-            patch.object(main_module, "AUTO_UPDATE_ENABLED", False),
-            patch("requests.get") as mock_get,
-            patch("subprocess.run") as mock_run,
+            patch.object(auto_update, "AUTO_UPDATE_ENABLED", False),
+            patch("validator.auto_update.requests.get") as mock_get,
+            patch("validator.auto_update.subprocess.run") as mock_run,
         ):
-            validator_instance._check_for_updates()
+            auto_update.check_for_updates()
             mock_get.assert_not_called()
             mock_run.assert_not_called()
 
-    def test_triggers_watchtower_waits_for_health_and_pulls(self, validator_instance):
+    def test_triggers_watchtower_waits_for_health_and_pulls(self):
         with (
-            patch.object(main_module, "AUTO_UPDATE_ENABLED", True),
-            patch("requests.get", side_effect=_mock_get()) as mock_get,
-            patch("subprocess.run", return_value=MagicMock(returncode=0)) as mock_run,
+            patch.object(auto_update, "AUTO_UPDATE_ENABLED", True),
+            patch(
+                "validator.auto_update.requests.get", side_effect=_mock_get()
+            ) as mock_get,
+            patch(
+                "validator.auto_update.subprocess.run",
+                return_value=MagicMock(returncode=0),
+            ) as mock_run,
         ):
-            validator_instance._check_for_updates()
+            auto_update.check_for_updates()
 
             urls = [c[0][0] for c in mock_get.call_args_list]
             assert any("/v1/update" in u for u in urls)
@@ -67,27 +65,32 @@ class TestCheckForUpdates:
             RuntimeError("unexpected"),
         ],
     )
-    def test_watchtower_errors_do_not_crash(self, validator_instance, watchtower_error):
+    def test_watchtower_errors_do_not_crash(self, watchtower_error):
         with (
-            patch.object(main_module, "AUTO_UPDATE_ENABLED", True),
+            patch.object(auto_update, "AUTO_UPDATE_ENABLED", True),
             patch(
-                "requests.get", side_effect=_mock_get(watchtower_resp=watchtower_error)
+                "validator.auto_update.requests.get",
+                side_effect=_mock_get(watchtower_resp=watchtower_error),
             ),
-            patch("subprocess.run", return_value=MagicMock(returncode=0)),
-        ):
-            validator_instance._check_for_updates()
-
-    def test_sandbox_pull_failure_does_not_crash(self, validator_instance):
-        with (
-            patch.object(main_module, "AUTO_UPDATE_ENABLED", True),
-            patch("requests.get", side_effect=_mock_get()),
             patch(
-                "subprocess.run", return_value=MagicMock(returncode=1, stderr="denied")
+                "validator.auto_update.subprocess.run",
+                return_value=MagicMock(returncode=0),
             ),
         ):
-            validator_instance._check_for_updates()
+            auto_update.check_for_updates()
 
-    def test_waits_for_health_after_watchtower(self, validator_instance):
+    def test_sandbox_pull_failure_does_not_crash(self):
+        with (
+            patch.object(auto_update, "AUTO_UPDATE_ENABLED", True),
+            patch("validator.auto_update.requests.get", side_effect=_mock_get()),
+            patch(
+                "validator.auto_update.subprocess.run",
+                return_value=MagicMock(returncode=1, stderr="denied"),
+            ),
+        ):
+            auto_update.check_for_updates()
+
+    def test_waits_for_health_after_watchtower(self):
         """Health check retries until proxy is ready."""
         health_calls = 0
 
@@ -103,11 +106,14 @@ class TestCheckForUpdates:
             return MagicMock(ok=True)
 
         with (
-            patch.object(main_module, "AUTO_UPDATE_ENABLED", True),
-            patch("requests.get", side_effect=side_effect),
-            patch("subprocess.run", return_value=MagicMock(returncode=0)),
-            patch("time.sleep"),
+            patch.object(auto_update, "AUTO_UPDATE_ENABLED", True),
+            patch("validator.auto_update.requests.get", side_effect=side_effect),
+            patch(
+                "validator.auto_update.subprocess.run",
+                return_value=MagicMock(returncode=0),
+            ),
+            patch("validator.auto_update.time.sleep"),
         ):
-            validator_instance._check_for_updates()
+            auto_update.check_for_updates()
 
         assert health_calls == 3
