+++ subnet/tests/validator/test_validate_inference_token.py
@@ -1,39 +1,22 @@
-"""Tests for Validator._validate_inference_token and _validation_model_for."""
+"""Tests for validator.inference_token (validate_inference_token, validation_model_for)."""
 
-import sys
-import types
 from unittest.mock import MagicMock, patch
 
 import pytest
 
-
-def _stub_bittensor_core():
-    """Stub out bittensor.core.subtensor so validator.main can be imported."""
-    if "bittensor.core.subtensor" not in sys.modules:
-        bt_core = sys.modules.setdefault(
-            "bittensor.core", types.ModuleType("bittensor.core")
-        )
-        subtensor_mod = types.ModuleType("bittensor.core.subtensor")
-        subtensor_mod.Subtensor = MagicMock()
-        sys.modules["bittensor.core.subtensor"] = subtensor_mod
-        bt_core.subtensor = subtensor_mod
-
-
-_stub_bittensor_core()
-
-from validator.main import Validator  # noqa: E402
+from validator.inference_token import validate_inference_token, validation_model_for
 
 
 class TestValidationModelFor:
     def test_chutes_returns_expected_model(self):
-        assert Validator._validation_model_for("chutes") == "Qwen/Qwen3-32B-TEE"
+        assert validation_model_for("chutes") == "Qwen/Qwen3-32B-TEE"
 
     def test_openrouter_returns_expected_model(self):
-        assert Validator._validation_model_for("openrouter") == "openai/gpt-oss-20b"
+        assert validation_model_for("openrouter") == "openai/gpt-oss-20b"
 
     def test_unknown_provider_raises(self):
         with pytest.raises(ValueError, match="unknown inference provider"):
-            Validator._validation_model_for("unknown_provider")
+            validation_model_for("unknown_provider")
 
 
 class TestValidateInferenceToken:
@@ -45,16 +28,22 @@ class TestValidateInferenceToken:
         return mock
 
     def test_200_returns_valid(self):
-        with patch("validator.main.requests.post", return_value=self._mock_resp(200)):
-            ok, reason = Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post",
+            return_value=self._mock_resp(200),
+        ):
+            ok, reason = validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         assert ok is True
         assert reason == ""
 
     def test_401_returns_invalid(self):
-        with patch("validator.main.requests.post", return_value=self._mock_resp(401)):
-            ok, reason = Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post",
+            return_value=self._mock_resp(401),
+        ):
+            ok, reason = validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         assert ok is False
@@ -63,26 +52,32 @@ class TestValidateInferenceToken:
     def test_402_returns_invalid_with_message(self):
         json_body = {"detail": {"message": "insufficient balance"}}
         with patch(
-            "validator.main.requests.post",
+            "validator.inference_token.requests.post",
             return_value=self._mock_resp(402, json_body),
         ):
-            ok, reason = Validator._validate_inference_token(
+            ok, reason = validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         assert ok is False
         assert "insufficient balance" in reason
 
     def test_429_returns_valid(self):
-        with patch("validator.main.requests.post", return_value=self._mock_resp(429)):
-            ok, reason = Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post",
+            return_value=self._mock_resp(429),
+        ):
+            ok, reason = validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         assert ok is True
         assert reason == ""
 
     def test_5xx_returns_valid(self):
-        with patch("validator.main.requests.post", return_value=self._mock_resp(503)):
-            ok, reason = Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post",
+            return_value=self._mock_resp(503),
+        ):
+            ok, reason = validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         assert ok is True
@@ -90,9 +85,10 @@ class TestValidateInferenceToken:
 
     def test_exception_returns_valid(self):
         with patch(
-            "validator.main.requests.post", side_effect=Exception("connection refused")
+            "validator.inference_token.requests.post",
+            side_effect=Exception("connection refused"),
         ):
-            ok, reason = Validator._validate_inference_token(
+            ok, reason = validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         assert ok is True
@@ -100,8 +96,10 @@ class TestValidateInferenceToken:
 
     def test_url_built_from_base_url(self):
         mock_resp = self._mock_resp(200)
-        with patch("validator.main.requests.post", return_value=mock_resp) as mock_post:
-            Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post", return_value=mock_resp
+        ) as mock_post:
+            validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1", "Qwen/Qwen3-32B-TEE"
             )
         args, kwargs = mock_post.call_args
@@ -109,8 +107,10 @@ class TestValidateInferenceToken:
 
     def test_url_strips_trailing_slash(self):
         mock_resp = self._mock_resp(200)
-        with patch("validator.main.requests.post", return_value=mock_resp) as mock_post:
-            Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post", return_value=mock_resp
+        ) as mock_post:
+            validate_inference_token(
                 "tok", "https://llm.chutes.ai/v1/", "Qwen/Qwen3-32B-TEE"
             )
         args, kwargs = mock_post.call_args
@@ -119,8 +119,10 @@ class TestValidateInferenceToken:
     def test_validate_inference_token_against_openrouter(self):
         """Verify the validator can smoke-test an OR-style endpoint."""
         mock_resp = self._mock_resp(200)
-        with patch("validator.main.requests.post", return_value=mock_resp) as mock_post:
-            ok, reason = Validator._validate_inference_token(
+        with patch(
+            "validator.inference_token.requests.post", return_value=mock_resp
+        ) as mock_post:
+            ok, reason = validate_inference_token(
                 "sk-or-test", "https://openrouter.ai/api/v1", "openai/gpt-oss-20b"
             )
